+++ lib/find-last.ts
@@ -1,10 +1,10 @@
 export default function findLast<T>(
-  arr: T[],
+  items: T[],
   callback: (element: T, index?: number, arr?: T[]) => boolean,
 ): T | undefined {
-  for (let index = arr.length - 1; index >= 0; index--) {
-    if (callback(arr[index], index, arr)) {
-      return arr[index];
+  for (let index = items.length - 1; index >= 0; index--) {
+    if (callback(items[index], index, items)) {
+      return items[index];
     }
   }