+++ lib/hex-to-rgb.ts
@@ -6,7 +6,7 @@ type RGB = {
 
 export default (hex: string): RGB | null => {
   const shorthandRegex = /^#?([a-f\d])([a-f\d])([a-f\d])$/i;
-  const longHex = hex.replace(shorthandRegex, (m, r, g, b) => r + r + g + g + b + b);
+  const longHex = hex.replace(shorthandRegex, (m: string, r: string, g: string, b: string) => r + r + g + g + b + b);
   const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(longHex);
 
   return result ? {